classes: simplify, allow appendage & sail change
Previously, hydrofoils were explicitly forbidden from being removed. As
not other appendages are mentioned, sails have been changed in
between competitions in the past, and a boat with functioning hydrofoils
would actually be quite impressive, the rule against changes was
removed.
Overall the description was generalised and simplified. Technically a
purely wave propelled boat may compete based on the simplified rules,
but again, it would be impressive to see this in competition with
sailing boats.